backend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
PyPDF2
camelot-py[cv]
pdfplumber
PyMuPDF
pandas
tiktoken
litellm
google-generativeai
openai
anthropic
